Moved create test to Integration test folder
